requirements.txt
absl-py==2.1.0
addict==2.4.0
albucore==0.0.17
albumentations==1.4.18
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.5.2
cachetools==5.5.0
certifi==2020.6.20
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.1.1
crcmod==1.7
cycler==0.12.1
Cython==3.2.4
diffusers==0.36.0
easydict==1.13
eval_type_backport==0.3.1
exceptiongroup==1.3.1
faster-coco-eval==1.7.0
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.35.0
grpcio==1.66.1
h11==0.16.0
h5py==3.11.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.6
jmespath==0.10.0
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
model-index==0.1.11
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pip @ file:///croot/pip_1723484598856/work
platformdirs==4.3.6
progress==1.6
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.0
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2023.4
PyWavelets==1.4.1
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1723018227672/work
regex==2024.11.6
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
rsa==4.9
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
sentry-sdk==2.48.0
setuptools==60.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2023.7.10
timm==1.0.24
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
typing_extensions==4.13.2
tzdata==2025.3
urllib3==1.26.20
wandb==0.23.1
wheel==0.44.0
zipp==3.20.2
